--- gsoc/cypress/e2e/jsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "lib": ["es5", "dom"],
-    "types": ["cypress"]
-  },
-  "include": ["./**/*", "../support/**/*"]
-}
