--- gsoc/package.json
@@ -1,44 +1,0 @@
-{
-  "name": "gsoc23",
-  "version": "0.0.0",
-  "private": true,
-  "module": "dist/vue.runtime.esm-bundler.js",
-  "scripts": {
-    "dev": "vite --open",
-    "build": "vite build",
-    "preview": "vite preview",
-    "test:unit": "jest",
-    "test:e2e": "start-server-and-test preview http://localhost:4173 'cypress run --e2e'",
-    "test:e2e:dev": "start-server-and-test 'vite dev --port 4173' http://localhost:4173 'cypress open --e2e'",
-    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore",
-    "format": "prettier --write src/",
-    "test": "jest --coverage"
-  },
-  "dependencies": {
-    "pinia": "^2.0.32",
-    "vue": "^2.6.14",
-    "vue-router": "^3.5.2"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.21.0",
-    "@babel/core": "^7.21.0",
-    "@babel/preset-env": "^7.20.2",
-    "@rushstack/eslint-patch": "^1.2.0",
-    "@vue/cli-plugin-unit-jest": "^5.0.8",
-    "@vue/eslint-config-prettier": "^7.1.0",
-    "@vue/test-utils": "^1.2.2",
-    "babel-jest": "^29.5.0",
-    "cypress": "^8.5.0",
-    "eslint": "^8.34.0",
-    "eslint-plugin-cypress": "^2.12.1",
-    "eslint-plugin-vue": "^9.9.0",
-    "jest": "^29.5.0",
-    "jsdom": "^21.1.0",
-    "prettier": "^2.8.4",
-    "start-server-and-test": "^2.0.0",
-    "vite": "^2.7.9",
-    "vitest": "^0.29.1",
-    "vue-jest": "^3.0.7",
-    "vue-template-compiler": "^2.6.14"
-  }
-}
