--- gsoc/src/components/ListItem.vue
@@ -1,16 +1,0 @@
-<template>
-  <li>
-    {{ item }}
-  </li>
-</template>
-
-<script>
-export default {
-  props: {
-    item: {
-      type: Object,
-      required: true
-    }
-  }
-};
-</script>
