--- gsoc/src/components/SugarizerFooter.vue
@@ -1,12 +1,0 @@
-<template>
-    <footer>
-      <p>&copy; 2023 Sugar Labs. All rights reserved.</p>
-    </footer>
-  </template>
-  
-  <script>
-  export default {
-    name: 'SugarizerFooter'
-  }
-  </script>
-
