--- gsoc/src/components/SugarizerHeader.vue
@@ -1,20 +1,0 @@
-<template>
-    <header>
-      <h1>Sugarizer VueJS Core</h1>
-      <nav>
-        <ul>
-          <li><a href="#">Home</a></li>
-          <li><a href="#">Login</a></li>
-          <li><a href="#">List View</a></li>
-          <li><a href="#">Settings</a></li>
-        </ul>
-      </nav>
-    </header>
-  </template>
-  
-  <script>
-  export default {
-    name: 'SugarizerHeader'
-  }
-  </script>
-
