--- gsoc/src/views/HomeView.vue
@@ -1,18 +1,0 @@
-<template>
-  <div class="home-view">
-    <h1>Home View</h1>
-    <p>Welcome to Sugarizer VueJS Core!</p>
-  </div>
-</template>
-
-<script>
-export default {
-  name: "HomeView",
-};
-</script>
-
-<style scoped>
-.home-view {
-  margin: 20px;
-}
-</style>
