+++ src/components/Doc/DocItem.jsx
@@ -1,14 +1,14 @@
 import { useState } from 'react';
+import { Link } from 'react-router-dom';
 
 
 const DocItem = ({ title, content }) => {
   const [readMore, setReadMore] = useState(false);
 
-  // const value = content.match(/{(.*?)}/g);
-  // var regex = new RegExp(value, "gi");
-  // const b = "<b>"
+  const value = content.match(/{(.*?)}/g);
+  var regex = new RegExp(value, "gi");
   return (
-    <div className={`p-4 rounded shadow-lg shadow-[#118d7c22] bg-white/5 backdrop-blur-[10px]`}>
+    <div className={`p-4 rounded shadow-lg shadow-[#118d7c20] bg-white/5 backdrop-blur-[10px]`}>
       <h6 className='font-semibold text-white text-opacity-80 mb-2'>{title}</h6>
       <p className={`text-sm text-gray line-clamp-11 leading-relaxed ${!readMore ? "none" : "hidden"}`}>
         {
@@ -22,7 +22,10 @@ const DocItem = ({ title, content }) => {
           <p key={index}>{content}</p>
         ))}
       </p>
-      {content.length > 100 ? <button onClick={() => { setReadMore(!readMore) }}>{readMore ? 'Hide' : 'Read'} </button> : ''}
+      {content.length > 100 ? <Link
+        to={title.split(" ").join("_").toLowerCase()}
+      // onClick={() => { setReadMore(!readMore) }}
+      >{readMore ? 'Hide' : 'Read'} </Link> : ''}
     </div>
   );
 };
