+++ src/data/error.json
@@ -12,6 +12,13 @@
       "description": "error: failed to push some refs to 'github.com:user/repo.git'",
       "solutions": "rm -rf .git<git init<git add .<git commit -m \"ges\"<git branch -M main<git remote add origin YOUR_GITHUB_REPO_LINK<git push -u origin main -f"
     },
+
+    {
+      "type": "merge",
+      "title": "fatal: refusing to merge unrelated histories",
+      "description": "Not able to merge in local and online.",
+      "solutions": "git merge majorUpdate --allow-unrelated-histories"
+    },
     {
       "type": "add",
       "title": "fatal: not a valid object name: 'master/main'",
