+++ src/main.jsx
@@ -5,10 +5,11 @@ import {
   createBrowserRouter,
   RouterProvider,
 } from "react-router-dom";
-import Doc from "./components/Doc/Doc";
 import Contributors from "./components/Contributors/contributors";
 import BGShape from "./components/BGShape";
 import NotFound from "./pages/404";
+import SingleDoc from "./pages/single doc";
+import Doc from "./pages/doc";
 /* Creating a router object that is used to render the correct component based on the url. */
 const router = createBrowserRouter(
   [
@@ -19,6 +20,10 @@ const router = createBrowserRouter(
     {
       path: '/doc',
       element: <Doc />
+    },
+    {
+      path: '/doc/:doc_name',
+      element: <SingleDoc />
     },
     {
       path: '/Contributors',
